Fix bug with hack
When moving a tile, I need to mark the location previously occupied by the
tile as free. However, I don't have easy access to the water object that is
at that location. The workaround is to use the water tile now covered by
the moving tile as the placeholder to mark that the tile is free.
Essentially, the two tiles swap location.
This works just fine now, but will likely lead to bugs.